+++ src/components/domain/Page/molecules/NoPageAlert/NoPageAlert.tsx
@@ -1,0 +1,18 @@
+import Image from 'next/image';
+import { VFC } from 'react';
+import { imagePath } from '~/const/imagePath';
+
+import { useLocale } from '~/hooks/useLocale';
+
+export const NoPageAlert: VFC = () => {
+  const { t } = useLocale();
+
+  return (
+    <div className="text-center">
+      <div className="w-25 mx-auto my-2">
+        <Image src={imagePath.NO_PAGE} height={958} width={1000} />
+      </div>
+      <h2 className="mb-0">{t['your_pages_don’t_exist']}</h2>
+    </div>
+  );
+};
