+++ src/components/domain/User/atoms/UserIcon/UserIcon.tsx
@@ -1,0 +1,13 @@
+import { VFC } from 'react';
+
+type Props = {
+  image: string;
+  size?: number;
+  isCircle?: boolean;
+};
+
+export const UserIcon: VFC<Props> = (props) => {
+  const { size = 32, isCircle = false } = props;
+
+  return <img height={size} width="auto" src={props.image} alt={props.image} className={`${isCircle ? 'rounded-circle' : ''}`} />;
+};
