+++ src/components/domain/User/atoms/UserIcon/index.tsx
@@ -1,13 +1,1 @@
-import { VFC } from 'react';
-
-type Props = {
-  image: string;
-  size?: number;
-  isCircle?: boolean;
-};
-
-export const UserIcon: VFC<Props> = (props) => {
-  const { size = 32, isCircle = false } = props;
-
-  return <img height={size} width="auto" src={props.image} className={`${isCircle ? 'rounded-circle' : ''}`} />;
-};
+export { UserIcon } from './UserIcon';
